requirements.txt
Flask>=2.0.0
Flask-WTF>=1.0.0
Flask-Login>=0.5.0
Flask-WeasyPrint>=0.5